pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "email-validator>=2.3.0",
    "pandas>=2.3.2",
    "pdfplumber>=0.11.7",
    "phonenumbers>=9.0.12",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.1.1",
    "python-slugify>=8.0.4",
    "requests>=2.32.5",
    "tenacity>=9.1.2",
]
